move common to shared
